refactor(views): Move view related files to view package
