+++ src/main/kotlin/lotto/views/InputView.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.views
 
 import camp.nextstep.edu.missionutils.Console
